+++ src/bin/resym.rs
@@ -198,7 +198,7 @@ impl<'p> ResymApp {
                     ) {
                         if let Err(err) = self
                             .backend
-                            .send_command(BackendCommand::LoadPDB(file_path))
+                            .send_command(BackendCommand::LoadPDB(file_path.into()))
                         {
                             log::error!("Failed to load the PDB file: {}", err);
                         } else {
@@ -239,13 +239,14 @@ impl<'p> ResymApp {
                                 .clicked()
                             {
                                 self.selected_row = row_index;
-                                let result =
-                                    self.backend.send_command(BackendCommand::ReconstructType(
+                                let result = self.backend.send_command(
+                                    BackendCommand::ReconstructTypeByIndex(
                                         *type_index,
                                         self.settings.print_header,
                                         self.settings.reconstruct_dependencies,
                                         self.settings.print_access_specifiers,
-                                    ));
+                                    ),
+                                );
                                 if let Err(err) = result {
                                     log::error!("Failed to reconstruct type: {}", err);
                                 }