+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Command-line version of the tool (`resymc`)
 
 ## [0.1.0] - 2022-05-04
 Initial release